DAS-1900: Added some unit tests and some error handling

+++ src/earthdataVarinfo/handler.py
@@ -1,4 +1,5 @@
 import os
+import traceback
 import json
 import shutil
 from tempfile import mkdtemp
@@ -18,41 +19,54 @@ def default(event, context):
     # print('Token in python lambda:'+token)
 
   # Get collectionConceptId from the query parameters
-  queryStringParameters = event['queryStringParameters']
-  if queryStringParameters != None and 'collection_concept_id' in queryStringParameters:
-    collectionConceptId = queryStringParameters['collection_concept_id']
+  query_string_parameters = event['queryStringParameters']
+  if query_string_parameters != None and 'collection_concept_id' in query_string_parameters:
+    collection_concept_id = query_string_parameters['collection_concept_id']
     # print('collection_concept_id:'+collectionConceptId)
 
-  # Get CMR environment from the cmrRootUrl
-  cmr_url = os.environ['cmrRootUrl']
-  cmr_env = 'PROD'
-  if ".sit." in cmr_url:
-    cmr_env = 'SIT'
-  if ".uat." in cmr_url:
-    cmr_env = 'UAT'
+  # Setup CMR search url from cmrRootUrl
+  cmr_url = os.environ['cmrRootUrl']+'/search/'
 
-  # Retrieve a list of 10 granules for the collection
-  granules = get_granules(collectionConceptId, cmr_env, token=token)
+  temp_dir = None
 
-  # Get the URL for the first granule (NetCDF-4 file):
-  granule_link = get_granule_link(granules)
-  print('Downloading granule from '+granule_link)
+  try:
+    # Retrieve a list of 10 granules for the collection
+    granules = get_granules(collection_concept_id, cmr_env=cmr_url, token=token)
 
-  # Make a temporary directory (this may cause permission issues?):
-  temp_dir = mkdtemp()
+    # Get the URL for the first granule (NetCDF-4 file):
+    granule_link = get_granule_link(granules)
+    print('Downloading granule from '+granule_link)
 
-  # Download file to lambda runtime environment
-  local_granule = download_granule(granule_link, token, out_directory=temp_dir)
-  print('Downloaded!')
+    # Make a temporary directory (this may cause permission issues?):
+    temp_dir = mkdtemp()
 
-  # Parse the granule with VarInfo:
-  var_info = VarInfoFromNetCDF4(local_granule)
+    # Download file to lambda runtime environment
+    local_granule = download_granule(granule_link, token, out_directory=temp_dir)
+    print('Downloaded!')
 
-  # Generate all the UMM-Var records:
-  all_variables = get_all_umm_var(var_info)
+    # Parse the granule with VarInfo:
+    var_info = VarInfoFromNetCDF4(local_granule)
 
-  # Delete temporary directory and files
-  shutil.rmtree(temp_dir)
+    # Generate all the UMM-Var records:
+    all_variables = get_all_umm_var(var_info)
+
+    # Delete temporary directory and files
+    shutil.rmtree(temp_dir)
+  except:
+    # Delete temporary directory and files if created
+    if temp_dir is not None:
+      shutil.rmtree(temp_dir)
+
+    # Print error message and stacktrace
+    print('Error executing earthdata-varinfo lambda')
+    tb = traceback.format_exc()
+    print(tb)
+    return {
+      'statusCode': 400,
+      'headers': {'Content-Type': 'application/json'},
+      'body': json.dumps({'Error': 'Error executing earthdata-varinfo lambda',
+                          'StackTrace': tb})
+  }
 
   # Return a successful response
   return {
